Move priority to dedicated module
