+++ src/inbox.rs
@@ -1,6 +1,7 @@
 //! Latency is prioritised over most accurate prioritisation. Specifically, at most one low priority
 //! message may be handled before piled-up higher priority messages will be handled.
 
+mod priority;
 pub mod rx;
 pub mod tx;
 mod waiting_receiver;
@@ -15,6 +16,7 @@ use std::task::{Context, Poll, Waker};
 use std::{cmp, mem};
 
 use event_listener::{Event, EventListener};
+pub use priority::{ByPriority, HasPriority, Priority};
 pub use rx::Receiver;
 pub use tx::Sender;
 
@@ -576,62 +578,5 @@ impl<A> Future for WaitingSender<A> {
             WaitingSender::Delivered => Poll::Ready(Ok(())),
             WaitingSender::Closed => Poll::Ready(Err(Error::Disconnected)),
         }
-    }
-}
-
-#[derive(Eq, PartialEq, Ord, PartialOrd, Copy, Clone)]
-pub enum Priority {
-    Valued(u32),
-    Shutdown,
-}
-
-impl Default for Priority {
-    fn default() -> Self {
-        Priority::Valued(0)
-    }
-}
-
-pub trait HasPriority {
-    fn priority(&self) -> Priority;
-}
-
-/// A wrapper struct that allows comparison and ordering for anything thas has a priority, i.e. implements [`HasPriority`].
-struct ByPriority<T>(pub T);
-
-impl<T> HasPriority for ByPriority<T>
-where
-    T: HasPriority,
-{
-    fn priority(&self) -> Priority {
-        self.0.priority()
-    }
-}
-
-impl<T> PartialEq for ByPriority<T>
-where
-    T: HasPriority,
-{
-    fn eq(&self, other: &Self) -> bool {
-        self.0.priority().eq(&other.0.priority())
-    }
-}
-
-impl<T> Eq for ByPriority<T> where T: HasPriority {}
-
-impl<T> PartialOrd for ByPriority<T>
-where
-    T: HasPriority,
-{
-    fn partial_cmp(&self, other: &Self) -> Option<Ordering> {
-        self.0.priority().partial_cmp(&other.0.priority())
-    }
-}
-
-impl<T> Ord for ByPriority<T>
-where
-    T: HasPriority,
-{
-    fn cmp(&self, other: &Self) -> Ordering {
-        self.0.priority().cmp(&other.0.priority())
     }
 }
